Fixed easyneff tests to use the public key correctly
Neff shuffling, at least as currently implemented in Kyber, depends on the
points to be shuffled being encrypted with the same public key as the H
parameter.

Also, as currently implemented in Kyber, the G parameter should be
nil (use the base point). So we remove it entirely from this implementation
to reduce code and confusion.

Also: Fix the cleartext for the tests to more accurately reflect
the DEDIS team's various college football preferences.